src/main/database/db.py
import os

from dotenv import load_dotenv

from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.declarative import declarative_base

load_dotenv()

engine = create_engine(url=os.getenv('SQLALCHEMY_DATABASE_URI'),
                       pool_size=10,
                       pool_pre_ping=True,
                       echo=True)
SessionLocal = sessionmaker(bind=engine, autoflush=False, autocommit=False)
Base = declarative_base()
