requirements.txt
blinker==1.8.2
Bcrypt-Flask==1.0.2
click==8.1.7
easy-setup==0.0.1
Flask[async]==3.0.3
Flask-JWT-Extended==4.6.0
gunicorn==22.0.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.8.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlalchemy==2.0.31
starlette==0.38.2
Werkzeug==3.0.3
uvicorn==0.30.3
zipp==3.19.2
